Decouple SDK from instrumentation
