+++ instrumentation/github.com/aws/aws-lambda-go/otellambda/go.mod
@@ -2,18 +2,10 @@ module go.opentelemetry.io/contrib/instrumentation/github.com/aws/aws-lambda-go/
 
 go 1.16
 
-replace (
-	go.opentelemetry.io/contrib/detectors/aws/lambda => ../../../../../detectors/aws/lambda
-	go.opentelemetry.io/contrib/propagators/aws => ../../../../../propagators/aws
-)
-
 require (
 	github.com/aws/aws-lambda-go v1.24.0
 	github.com/stretchr/testify v1.7.0
-	go.opentelemetry.io/contrib/detectors/aws/lambda v0.21.0
 	go.opentelemetry.io/contrib/propagators/aws v0.21.0
 	go.opentelemetry.io/otel v1.0.0-RC1
-	go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracegrpc v1.0.0-RC1
-	go.opentelemetry.io/otel/sdk v1.0.0-RC1
 	go.opentelemetry.io/otel/trace v1.0.0-RC1
 )
