+++ instrumentation/github.com/aws/aws-lambda-go/otellambda/test/go.mod
@@ -1,0 +1,19 @@
+module go.opentelemetry.io/contrib/instrumentation/github.com/aws/aws-lambda-go/otellambda/test
+
+go 1.16
+
+replace (
+	go.opentelemetry.io/contrib/detectors/aws/lambda => ../../../../../../detectors/aws/lambda
+	go.opentelemetry.io/contrib/instrumentation/github.com/aws/aws-lambda-go/otellambda => ../
+)
+
+require (
+	github.com/aws/aws-lambda-go v1.25.0
+	github.com/stretchr/testify v1.7.0
+	go.opentelemetry.io/contrib/detectors/aws/lambda v0.21.0
+	go.opentelemetry.io/contrib/instrumentation/github.com/aws/aws-lambda-go/otellambda v0.21.0
+	go.opentelemetry.io/contrib/propagators/aws v0.21.0
+	go.opentelemetry.io/otel v1.0.0-RC1
+	go.opentelemetry.io/otel/sdk v1.0.0-RC1
+	go.opentelemetry.io/otel/trace v1.0.0-RC1
+)
